% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{remove_non_phyto}
\alias{remove_non_phyto}
\title{Remove Non-Phytoplankton Taxa from a Dataset}
\usage{
remove_non_phyto(df)
}
\arguments{
\item{df}{A dataframe with a \code{Taxon} column}

\item{exclude_file}{Path to a CSV file with a single column named \code{Taxon}
containing taxa to be excluded}
}
\value{
The filtered dataframe, with an attribute \code{'log'} containing
a list named \code{non_phyto_removed}, including a character vector of
removed taxa
}
\description{
Removes rows from a dataframe where the \code{Taxon} matches an entry in an external
CSV file of known non-phytoplankton taxa. Returns the filtered dataframe and logs
the number and names of removed taxa.
}
